rename glist to dlnode (#117)
